requirements/common.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
marshmallow-sqlalchemy
alembic
bleach
blinker
click
dominate
Flask-Bootstrap
Flask-HTTPAuth
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Principal
Flask-Moment
Flask-PageDown
Flask-WTF
html5lib
itsdangerous
Jinja2
Mako
Markdown
MarkupSafe
python-dateutil
python-dotenv
python-editor
six
visitor
webencodings
WTForms
email_validator
flask-change-password
httpx
psycopg2-binary
gunicorn
Flask-OAuthlib
Flask-RestCountries
pycountry
firebase-admin
Werkzeug==2.2.2
qrcode[pil]
fpdf
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
newsdataapi
Faker
Flask-Babel
opencage
firebase-admin
python-barcode
twilio
pyppeteer
WeasyPrint
apscheduler
reportlab
openpyxl
imapclient
pyzmail36
cairocffi